Move SGLang tests from mi300x-3 to mi300x-4
